third_party/blink/web_tests/platform/generic/fast/dom/Window/HTMLFrameSetElement-window-eventListener-attributes-expected.txt

This tests that setting window event listeners on the frameset, sets them on the window.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.onblur is func
PASS window.onblur is frameSet.onblur
PASS window.onfocus is func
PASS window.onfocus is frameSet.onfocus
PASS window.onerror is func
PASS window.onerror is frameSet.onerror
PASS window.onload is func
PASS window.onload is frameSet.onload
PASS window.onbeforeunload is func
PASS window.onbeforeunload is frameSet.onbeforeunload
PASS window.onhashchange is func
PASS window.onhashchange is frameSet.onhashchange
PASS window.onmessage is func
PASS window.onmessage is frameSet.onmessage
PASS window.onoffline is func
PASS window.onoffline is frameSet.onoffline
PASS window.ononline is func
PASS window.ononline is frameSet.ononline
PASS window.onresize is func
PASS window.onresize is frameSet.onresize
PASS window.onscroll is func
PASS window.onscroll is frameSet.onscroll
PASS window.onstorage is func
PASS window.onstorage is frameSet.onstorage
PASS window.onunload is func
PASS window.onunload is frameSet.onunload
PASS successfullyParsed is true

TEST COMPLETE
